requirements.txt
Imath==0.0.2
matplotlib==3.8.0
natsort==8.4.0
numpy==1.24.3
opencv_python==4.8.0.76
opencv_python_headless==4.8.0.76
openexr-python>=1.3.8
prettytable==3.10.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.0
tensorboard==2.14.0